move all the methods and fields to appropriate LLMChatThreadHeaderView model
