+++ llm_thread/static/src/models/llm_chat_view.js
@@ -63,7 +63,8 @@ registerModel({
                 }
                 return {
                     threadViewer: this.threadViewer,
-                    messageListView: {}, // This is crucial - we create MessageListView here
+                    messageListView: {},
+                    llmChatThreadHeaderView: {},
                 };
             },
         }),
