+++ llm_thread/static/src/models/main.js
@@ -10,3 +10,4 @@ import '@llm_thread/models/composer_view';
 import '@llm_thread/models/llm_model';
 import '@llm_thread/models/llm_provider';
 import '@llm_thread/models/thread_view';
+import '@llm_thread/models/llm_chat_thread_header_view';
